@@ -80,18 +80,18 @@ public class LocationsExtractor implements EntityDocumentDumpProcessor {
 					for (Statement statement : statements) {
 
 						if (statement.getClaim() != null && statement.getClaim().getMainSnak() != null
-								&& statement.getClaim().getMainSnak().getValue() != null) {
+								&& statement.getClaim().getValue() != null) {
 
 							String id = null;
 							try {
-								id = ((ItemIdValue) statement.getClaim().getMainSnak().getValue()).getId();
+								id = ((ItemIdValue) statement.getClaim().getValue()).getId();
 							} catch (ClassCastException e) {
 								continue;
 							}
 
 							if (id != null) {
 								this.itemsWithLocationCount++;
-								outLocations.print(itemDocument.getItemId().getId());
+								outLocations.print(itemDocument.getEntityId().getId());
 								outLocations.print(Config.TAB);
 								outLocations.print(csvEscape(itemDocument.findLabel("en")));
 								outLocations.print(Config.TAB);
